Directory reorganization
Reorganizing the directories to seperate out server and client
