+++ .elasticbeanstalk/.ebignore
@@ -1,7 +1,7 @@
-../node_modules/
+../server/node_modules/
 ../client/node_modules/
-../config/production.json
+../server/config/production.json
 ../client/build
 ../client/.eslintcache
-../package-lock.json
+../server/package-lock.json
 ../client/package-lock.json
