+++ server/config/courses.json
