+++ server/routes/api/profile.js
@@ -3,7 +3,7 @@ const request = require("request");
 const config = require("config");
 const router = express.Router();
 const auth = require("../../middleware/auth");
-const partialMatch = require("../../utils/utilities");
+const partialMatch = require("../../../utils/utilities");
 const { check, validationResult } = require("express-validator");
 const mongoose = require("mongoose");
 const Profile = require("../../models/Profile");
