+++ server/routes/api/searchbar.js
@@ -6,7 +6,7 @@ const router = express.Router();
 // @access Pubic
 router.get('/suggestedlist', (req,res) => {
     var spawn = require('child_process').spawn;
-    const process = spawn('python3', ['./algos/SearchBar/Trie.py', req.query.name]);
+    const process = spawn('python3', ['../../../algos/SearchBar/Trie.py', req.query.name]);
     process.stdout.on('data', (data) => {
         res.send(data.toString());
     });
